feat: split select into a control and field composables
