+++ packages/core/src/useSelect/index.ts
@@ -1,1 +1,2 @@
 export * from './useSelect';
+export * from './useSelectControl';
